refactored and restored scrolling after email template update
